chore: implemt `ExecutionContext`
